Refactor, rename, update tests
