+++ lib/ingram_micro/base_element.rb
