+++ lib/ingram_micro/elements/detail.rb
@@ -9,7 +9,7 @@ class IngramMicro::Detail < IngramMicro::BaseElement
   end
 
   def build(builder)
-    element[:line_items].each_with_index do |line_item, idx|
+    @element[:line_items].each_with_index do |line_item, idx|
       line_item.line_no = idx + 1 if line_item.line_no.nil?
       puts line_item.line_no
       builder.send("line-item") do
